delete_node() modified to delete gridfs object only when it's not present in multiple objects

+++ gnowsys-ndf/gnowsys_ndf/ndf/views/methods.py
@@ -3940,9 +3940,10 @@ def delete_node(
                     # print "\n 10 >> node found as File; nodes in GridFS : ", len(node_to_be_deleted.fs_file_ids)
                     if node_to_be_deleted.fs_file_ids:
                         for each in node_to_be_deleted.fs_file_ids:
-                            if node_to_be_deleted.fs.files.exists(each):
+                            if node_to_be_deleted.fs.files.exists(each) and node_collection.find({'fs_file_ids': {'$in': [each]} }).count() == 1:
                                 # print "\tdeleting node in GridFS : ", each
                                 node_to_be_deleted.fs.files.delete(each)
+
 
                 # Finally delete the node
                 node_to_be_deleted.delete()
